Working driver (#16)
* Initial commit, parse_urc fails

* Test pass, Response parse not working..

* Maked unrechable code

* Progress on custom digest

* Passing tests for custom digest

* Check for funky behaviour in tests

* Edm at no longer ref and various fixes

* change commands to edm

* Switch to EDM on startup

* Updating commands

* Updating EDM, fixing parse, tests for parse, fix SwitchToEDM command, handling startup message

* Handle EDM startup, handle AT startup and trim URC messages

* State control for reciving

* First working connect to WiFi

* Checkpoint 1

* New type for Bssid

* New Edm structure and structs for data transfer and events

* cleanup

* Channel_id and endpoint for all sockets

* get socket type by endpoint and channel_id

* TCP stack + event handeling in clinet + fixes for data formats

* Dummy DNS

* spin errors and handle all URCs each spin

* Network and WiFi error types

* Fixed Response handeling and parsing

* New URL building scheme..

* Updated digest and packet size + type correction

* TCP state is now SynSent before connection is established

* New error types and corrected spin method

* shrunk EdmDataEvent command lenght

* Cleanup of EDM wrapper

* Ip address parsing temp fix

* Import certificates working

* Working TLS

* Ping

* Working DNS

* Checkpoint

* Closure based URC handeling, fixed closing sockets, DNS spin error fix plus various other fixes

* Logging now defmt

* Defmt fixes and compatability with factbird firmware

* Handle double network up before Network working

* is_connected() and blocking TCP connect() + defmt changes

* rm --cached Crago.lock

* TLS enabled universially by importing credentials

* TLS single config

* Checkpoint

* defmt for error

* Working UDP implementation

* Changes to sockts, comments, logging and error

* defmt debug fixes

* working with new atat

* Fix serde usage and initial dependency cleanup

* fmt

* Bump embedded-nal to 0.2.0

* Fixed connect error handeling

* fix

* Fix comments + fmt + add unimplemented error

* Fixing some warnings

* Fixing compiler warnings

* New ReadMe etc.

* edition fix...

* Fix tests

* github actions

* Remove dev-dependencies

* fmt

* Handle module is in EDM mode on startup

* Revert last commit - fix not working

* Formatting

+++ ublox-short-range/src/command/data_mode/mod.rs
@@ -1,10 +1,9 @@
-//! ### 5 - Data Mode 
+//! ### 5 - Data Mode
 pub mod responses;
 pub mod types;
 pub mod urc;
 
 use atat::atat_derive::AtatCmd;
-use heapless::{consts, String};
 use responses::*;
 use types::*;
 
@@ -16,7 +15,7 @@ use super::NoResponse;
 /// After executing the data mode command or the extended data mode command, a delay of 50 ms is
 /// required before start of data transmission.
 #[derive(Clone, AtatCmd)]
-#[at_cmd("O", NoResponse, timeout_ms = 10000)]
+#[at_cmd("O", NoResponse, timeout_ms = 10000, value_sep = false)]
 pub struct ChangeMode {
     #[at_arg(position = 0)]
     pub mode: Mode,
@@ -29,7 +28,7 @@ pub struct ChangeMode {
 /// event.
 #[derive(Clone, AtatCmd)]
 #[at_cmd("+UDCP", ConnectPeerResponse, timeout_ms = 10000)]
-pub struct ConnectPeer <'a> {
+pub struct ConnectPeer<'a> {
     #[at_arg(position = 0, len = 128)]
     pub url: &'a str,
 }
@@ -41,7 +40,7 @@ pub struct ConnectPeer <'a> {
 #[at_cmd("+UDCPC", NoResponse, timeout_ms = 10000)]
 pub struct ClosePeerConnection {
     #[at_arg(position = 0)]
-    pub peer_handle: u32,
+    pub peer_handle: usize,
 }
 
 /// 5.4 Default remote peer +UDDRP
